requirements.txt
# Core runtime dependencies
google-generativeai>=0.8.5,<1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
jinja2>=3.1.0,<4.0.0
tqdm>=4.66.0,<5.0.0PyPDF2>=3.0.0,<4.0.0
PyPDF2>=3.0.0,<4.0.0
PyPDF2>=3.0.0,<4.0.0
